crypto(hw): fix byte/word-stride mismatch in AES-NI/SHA-NI memory access (#1223)
Real x86_64 hardware verification (hl-master) crashed with "panic: index out
of range" on the first hardware-path AES call — root cause: Bit's dynamic
[]T slice stores every element as one FULL 8-BYTE WORD regardless of T's own
width (ABI.md §2.1), not packed C-style bytes. cryptohw.zig's loadBlock/
storeBlock and the SHA-256 state/block accessors treated ptrOf([]byte) and
ptrOf([]u32) as packed contiguous arrays, reading/writing 16 (or 4) real
bytes per logical block instead of gathering/scattering across the true
8-byte-per-element stride — silently corrupting 8x the intended memory past
every buffer.

Fixed by threading every AES-NI/SHA-NI memory access through wordAt/wordAtMut
(byte address = base + logical_index*8) instead of native pointer-plus-index
arithmetic. GHASH was never affected: gcmMulHWords already passes plain u64
scalars (word-sized = no padding gap for u64 elements either).

Caught because arm64 never exercises this code (hwAvailable() is always
false there) and the local x86_64 target only cross-compiles, never runs —
this class of bug is exactly why the task's real-hardware verification step
exists, not a formality.

(cherry picked from commit e86f55b7cff5a340b1555160496322345f8de373)

+++ runtime/root.zig
@@ -1495,7 +1495,7 @@ export fn bit_rt_crypto_ghash_mul_hw(acc0: u64, acc1: u64, b0: u64, b1: u64, h0:
 /// FIPS-180-4 round function `sha256.bit`'s software `compress` runs (shared
 /// by SHA-224 and SHA-256, which differ only in IV/output truncation, both
 /// applied outside `compress`).
-export fn bit_rt_crypto_sha256_compress_hw(state: [*]u32, block: [*]const u8) callconv(.c) void {
+export fn bit_rt_crypto_sha256_compress_hw(state: [*]u8, block: [*]const u8) callconv(.c) void {
     if (is_x86_64) cryptohw.compress(state, block) else @panic("bit_rt_crypto_sha256_compress_hw: x86-64 only");
 }
